xml carender, myinformation Activity
